perf(generate): shrink `Symbol::index` from `usize` to `u32`
This cuts `Symbol`'s size from 16 to 8 bytes. Results in a ~2-7% wall time
reduction, and ~7-10% peak RSS reduction.

+++ crates/generate/src/prepare_grammar/intern_symbols.rs
@@ -144,7 +144,7 @@ pub(super) fn intern_symbols(
 
     for s in &supertypes {
         if s.is_non_terminal() {
-            kinds[s.index] = VariableType::Hidden;
+            kinds[s.index as usize] = VariableType::Hidden;
         }
     }
 
@@ -177,7 +177,7 @@ fn intern_root(
                     id,
                     Rule::Sym {
                         kind: s.kind,
-                        index: s.index as u32,
+                        index: s.index,
                     },
                 ),
                 None => Err(InternSymbolsError::Undefined(pool.resolve(sid).to_string()))?,
